sockets in a separate module #228
